% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_color_fill.R
\name{branded_pal}
\alias{branded_pal}
\title{Branded colour palette}
\usage{
branded_pal(primary = "blue", other = "red", direction = 1)
}
\arguments{
\item{primary}{Primary colour in palette.}

\item{other}{Secondary colour.}

\item{direction}{Direction of colour increase.}
}
\description{
Branded colour palette
}
\references{
\itemize{
 \item{https://www.garrickadenbuie.com/blog/custom-discrete-color-scales-for-ggplot2/}
}
}
