% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_check_input.R
\name{check_input}
\alias{check_input}
\title{Check input prediction and observation for possible errors}
\usage{
check_input(
  y_pred,
  y_pred_threshold,
  y_pred_bin,
  y_obs,
  y_pos,
  evaluate_calibration
)
}
\arguments{
\item{y_pred}{Predicted risk (or score), if available. Use \code{y_pred_bin}
if only predicted class label is available.}

\item{y_pred_threshold}{Threshold for \code{y_pred}.
\code{y_pred > y_pred_threshold} will be considered positive.}

\item{y_pred_bin}{Predicted class label. Ignored if \code{y_pred} is provided.}

\item{y_obs}{Observed class label.}

\item{y_pos}{Character representing positive class. Default is "1" for 0/1
encoding.}
}
\value{
Returns a list of cleaned \code{y_pred_threshold} and \code{y_obs}
  (both using 0/1 encoding).
}
\description{
Check input prediction and observation for possible errors
}
