% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_check_input.R
\name{check_pred_bin}
\alias{check_pred_bin}
\title{Check observed binary outcome}
\usage{
check_pred_bin(y_pred, y_pos, y_obs)
}
\arguments{
\item{y_pred}{A vector of predicted outcome labels. Should not contain
missing value.}

\item{y_pos}{A character representing the positive class in \code{y_obs}. If
\code{y_obs} is a factor, by default \code{y_pos = levels(y_obs)[2]}. If
\code{y_obs} is numeric or character, by default \code{y_pos = "1"} for 0/1
encoding.}

\item{y_obs}{A vector of observed binary outcome. Can be numeric, character
or factor. Should not contain missing value.}
}
\description{
Check observed binary outcome
}
