% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_check_input.R
\name{check_pred_general}
\alias{check_pred_general}
\title{Check prediction (general)}
\usage{
check_pred_general(y_pred, y_obs)
}
\arguments{
\item{y_pred}{A numeric vector of predicted probabilities. Should not contain
missing value.}

\item{y_obs}{A vector of observed binary outcome. Can be numeric, character
or factor. Should not contain missing value.}
}
\description{
Check prediction (general)
}
