% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_check_input.R
\name{check_pred_prob}
\alias{check_pred_prob}
\title{Check predicted probability}
\usage{
check_pred_prob(y_pred, y_pred_threshold, y_obs)
}
\arguments{
\item{y_pred}{A numeric vector of predicted probabilities. Should not contain
missing value.}

\item{y_pred_threshold}{Threshold for \code{y_pred}. Predict positive label
if \code{y_pred > y_pred_threshold}. Default is \code{NULL}, in which case
the threshold will be selected based on the ROC curve for \code{y_pred}.}

\item{y_obs}{A vector of observed binary outcome. Can be numeric, character
or factor. Should not contain missing value.}
}
\description{
Check predicted probability
}
