% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_check_input.R
\name{check_sens_ref}
\alias{check_sens_ref}
\title{<Private function> Check reference levels for sensitive variables}
\usage{
check_sens_ref(sens_var_df, sens_var_ref)
}
\arguments{
\item{sens_var_df}{A data.frame of sensitive variable(s).}

\item{sens_var_ref}{Reference class(es) of sensitive variable(s). Default is the
default reference category in \code{sens_var}.}
}
\description{
<Private function> Check reference levels for sensitive variables
}
