% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_check_input.R
\name{check_sens_var}
\alias{check_sens_var}
\title{Check and process sensitive variable}
\usage{
check_sens_var(sens_var, sens_var_ref)
}
\arguments{
\item{sens_var}{Sensitive variable(s). Must be categorical.}

\item{sens_var_ref}{Reference class(es) of sensitive variable(s). Default is the
default reference category in \code{sens_var}.}
}
\value{
Returns a vector of processed sensitive variable as a factor with
  specified reference class.
}
\description{
Check and process sensitive variable
}
