% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{compas}
\alias{compas}
\title{COMPAS data for demonstration}
\format{
An object of class \code{data.frame} with 6172 rows and 7 columns.
}
\usage{
compas
}
\description{
From fairmodel package
}
\keyword{datasets}
