% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_metrics.R
\name{compute_ci_boot}
\alias{compute_ci_boot}
\title{Computes selected performance metrics and bootstrap confidence interval (CI)}
\usage{
compute_ci_boot(y_pred_bin, y_obs, y_pos, B, conf_level)
}
\arguments{
\item{y_pred_bin}{Predicted class label.}

\item{y_obs}{Observed class label.}

\item{y_pos}{Character representing positive class. Default is "1" for 0/1
encoding.}

\item{B}{Number of bootstrap samples generated with replacement. Default is 1000.}

\item{conf_level}{Confidence level of the CI. Default is 0.95.}
}
\value{
Returns a data.frame generated from \code{compute_metrics()}, with
  two additional columns \code{lower} and \code{upper} of the bootstrap CI
  derived based on the quartiles of the bootstrap samples.
}
\description{
Computes selected performance metrics and bootstrap confidence interval (CI)
}
