% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_metrics.R
\name{compute_ci_cp}
\alias{compute_ci_cp}
\title{Compute performance metrics with Clopper–Pearson 95\% CI}
\usage{
compute_ci_cp(numerator, denominator, alpha = 0.05)
}
\arguments{
\item{numerator}{Numerator for performance metrics formula.}

\item{denominator}{Denominator for performance metrics formula.}

\item{alpha}{Target error rate. Default is 0.05 for 95\% CI.}
}
\value{
Returns a data.frame of the performance metrics (\code{est}) and the
  95\% CI (\code{lower}, \code{upper}). Consistent with the 95\% CI of
  accuracy returned from \code{carret::confusionMatrix} and
  \code{epiR::epi.tests} (default setting).
}
\description{
Compute performance metrics with Clopper–Pearson 95\% CI
}
\references{
\itemize{
 \item{Clopper CJ, Pearson ES. The use of confidence or fiducial limits illustrated in the case of the binomial. Biometrika. 1934 Dec 1;26(4):404-13.}
}
}
