% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_metrics.R
\name{compute_metrics}
\alias{compute_metrics}
\title{Compute selected performance metrics}
\usage{
compute_metrics(y_pred_bin, y_obs, y_pos = "1")
}
\arguments{
\item{y_pred_bin}{Predicted class label.}

\item{y_obs}{Observed class label.}

\item{y_pos}{Character representing positive class. Default is "1" for 0/1
encoding.}
}
\value{
Returns a data.frame of the names (\code{metric}) and estimated
  values (\code{est}) of the performance metrics.
}
\description{
Compute selected performance metrics
}
