% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_metrics.R
\name{eval_calib}
\alias{eval_calib}
\title{Compute calibration curves}
\usage{
eval_calib(y_pred, y_obs, sens_var)
}
\arguments{
\item{y_pred}{Processed predicted probability or score.}

\item{y_obs}{Processed observed label.}

\item{sens_var}{Sensitive variable.}
}
\value{
Returns a data.frame of data for calibration curves.
}
\description{
Compute calibration curves
}
