% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_metrics.R
\name{eval_metrics_based}
\alias{eval_metrics_based}
\title{Evaluate performance-based metrics}
\usage{
eval_metrics_based(y_pred_bin, y_obs, sens_var)
}
\arguments{
\item{y_pred_bin}{Processed predicted label.}

\item{y_obs}{Processed observed label.}

\item{sens_var}{Sensitive variable.}
}
\value{
Returns a list of \code{df_prob} (predicted vs observed probability)
  and \code{df_metrics} (predictive performance).
}
\description{
Evaluate performance-based metrics
}
