% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_metrics.R
\name{eval_metrics_group}
\alias{eval_metrics_group}
\title{Evaluate model performance at different thresholds by group}
\usage{
eval_metrics_group(y_pred, y_obs, sens_var)
}
\arguments{
\item{y_pred}{Processed predicted probabilities or scores.}

\item{y_obs}{Processed observed label.}

\item{sens_var}{Sensitive variable.}
}
\description{
Evaluate model performance at different thresholds by group
}
