% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_metrics.R
\name{eval_pred}
\alias{eval_pred}
\title{Evaluates fairness metrics within a subject group}
\usage{
eval_pred(y_pred_bin, y_obs, y_pos = "1")
}
\arguments{
\item{y_pred_bin}{Predicted class label.}

\item{y_obs}{Observed class label.}

\item{y_pos}{Character representing positive class. Default is "1" for 0/1
encoding.}
}
\value{
Returns a data.frame of metrics with Clopper–Pearson 95\% CI
}
\description{
Evaluates fairness metrics within a subject group
}
