% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_metrics.R
\name{eval_roc}
\alias{eval_roc}
\title{Compute ROC curves}
\usage{
eval_roc(y_pred, y_obs, sens_var)
}
\arguments{
\item{y_pred}{Processed predicted probability.}

\item{y_obs}{Processed observed label.}

\item{sens_var}{Sensitive variable.}
}
\value{
Returns a list containing a data.frame of data for ROC curves and a
  data.frame of AUC and 95\% DeLong CI.
}
\description{
Compute ROC curves
}
