% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_metrics.R
\name{evaluate_prediction}
\alias{evaluate_prediction}
\alias{plot.seeBias}
\title{Evaluates fairness metrics based on sensitive variable(s)}
\usage{
evaluate_prediction(
  y_pred = NULL,
  y_pred_threshold = NULL,
  y_pred_bin = NULL,
  y_obs,
  y_pos = "1",
  sens_var,
  sens_var_ref = NULL,
  evaluate_calibration = TRUE
)

\method{plot}{seeBias}(
  x,
  print_statistics = TRUE,
  labels = c("A", "B", "C", "D", "E"),
  heights = c(1.8, 1, 1),
  y = NULL,
  ...
)
}
\arguments{
\item{y_pred}{Predicted risk (or score), if available. Use \code{y_pred_bin}
if only predicted class label is available.}

\item{y_pred_threshold}{Threshold for \code{y_pred}.
\code{y_pred > y_pred_threshold} will be considered positive.}

\item{y_pred_bin}{Predicted class label. Ignored if \code{y_pred} is provided.}

\item{y_obs}{Observed class label.}

\item{y_pos}{Character representing positive class. Default is "1" for 0/1
encoding.}

\item{sens_var}{Sensitive variable(s). Must be categorical.}

\item{sens_var_ref}{Reference class(es) of sensitive variable(s). Default is the
default reference category in \code{sens_var}.}

\item{x}{\code{FairLite} object to plot}

\item{labels}{Labels for the 3 sub-figures in the combine figure. Default is
A, B and C.}

\item{heights}{Relative height of metrics figure (top) and the two figures
(bottom). Default is 2:1.}

\item{y}{Not supported.}

\item{...}{Not supported.}
}
\value{
Returns a data.frame of performance metrics evaluated within each
  sensitive group.
}
\description{
Evaluates fairness metrics based on sensitive variable(s)
}
\section{Functions}{
\itemize{
\item \code{plot(seeBias)}: Plot fairness metrics

}}
