% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_metrics.R
\name{find_y_pred_threshold}
\alias{find_y_pred_threshold}
\title{Find best threshold for predicted score from ROC analysis}
\usage{
find_y_pred_threshold(y_pred, y_obs, y_pos)
}
\arguments{
\item{y_obs}{Observed class label.}

\item{y_pos}{Character representing positive class. Default is "1" for 0/1
encoding.}
}
\value{
Returns the best threshold.
}
\description{
Find best threshold for predicted score from ROC analysis
}
