% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_color_fill.R
\name{pal_jama_fill}
\alias{pal_jama_fill}
\alias{scale_fill_jamma_alpha}
\title{Modified fill palette baesd on ggsci::pal_jama}
\usage{
pal_jama_fill(palette = c("default"), alpha = 1)

scale_fill_jamma_alpha(palette = c("default"), alpha = 1, ...)
}
\description{
Modified fill palette baesd on ggsci::pal_jama
}
\section{Functions}{
\itemize{
\item \code{scale_fill_jamma_alpha()}: Fill scale based on pal_jama_fill

}}
