% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_plot.R
\name{plot_calib_large}
\alias{plot_calib_large}
\title{Private function to plot calibration in the large}
\usage{
plot_calib_large(x)
}
\arguments{
\item{x}{\code{seeBias} object}
}
\description{
Private function to plot calibration in the large
}
