% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_plot.R
\name{plot_calibration}
\alias{plot_calibration}
\title{Private function to plot calibration curves}
\usage{
plot_calibration(x, print_statistics)
}
\arguments{
\item{x}{\code{FairLite} object to plot}
}
\description{
Private function to plot calibration curves
}
