% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_plot.R
\name{plot_metrics}
\alias{plot_metrics}
\title{Private function to plot fairness metrics}
\usage{
plot_metrics(x)
}
\arguments{
\item{x}{\code{FairLite} object to plot}
}
\description{
Private function to plot fairness metrics
}
