% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_plot.R
\name{plot_metrics_group}
\alias{plot_metrics_group}
\title{Private function to plot numbers needed by group}
\usage{
plot_metrics_group(x)
}
\arguments{
\item{x}{\code{seeBias} object}
}
\description{
Private function to plot numbers needed by group
}
