% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_plot.R
\name{plot_score}
\alias{plot_score}
\title{Private function to plot score distributions by label and group}
\usage{
plot_score(x)
}
\arguments{
\item{x}{\code{seeBias} object}
}
\description{
Private function to plot score distributions by label and group
}
