% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_color_fill.R
\name{scale_colour_branded}
\alias{scale_colour_branded}
\alias{scale_color_branded}
\alias{scale_fill_branded}
\title{Define scale function for colour}
\usage{
scale_colour_branded(primary = "blue", other = "red", direction = 1, ...)

scale_color_branded(primary = "blue", other = "red", direction = 1, ...)

scale_fill_branded(primary = "blue", other = "red", direction = 1, ...)
}
\arguments{
\item{primary}{Primary colour in palette.}

\item{other}{Secondary colour.}

\item{direction}{Direction of colour increase.}

\item{...}{Other parameters passed to \code{ggplot2::discrete_scale}.}
}
\description{
Define scale function for colour
}
\section{Functions}{
\itemize{
\item \code{scale_color_branded()}: Alias to scale_colour_branded

\item \code{scale_fill_branded()}: Define scale function for fill

}}
