% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_color_fill.R
\name{select_scale}
\alias{select_scale}
\title{Select appropriate color/fill scale based on number of sensitive groups}
\usage{
select_scale(x, type = "color")
}
\arguments{
\item{x}{seeBias object.}

\item{type}{Type to return, either "color" or "fill".}
}
\description{
Select appropriate color/fill scale based on number of sensitive groups
}
