% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_check_input.R
\name{sens_var_to_df}
\alias{sens_var_to_df}
\title{<Private function> Convert sensitive variable(s) to a data.frame of vector(s)}
\usage{
sens_var_to_df(sens_var)
}
\arguments{
\item{sens_var}{Sensitive variable(s). Must be categorical.}
}
\description{
<Private function> Convert sensitive variable(s) to a data.frame of vector(s)
}
