% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_check_input.R
\name{configure_group}
\alias{configure_group}
\title{Configure sensitive group}
\usage{
configure_group(group, ref, ref_label = NULL)
}
\arguments{
\item{group}{Original group variable.}

\item{ref}{Reference class of sensitive variable.}
}
\description{
Configure sensitive group
}
