% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_plot.R
\name{plot_calibration}
\alias{plot_calibration}
\title{Private function to plot calibration curves}
\usage{
plot_calibration(x, print_statistics)
}
\arguments{
\item{x}{\code{seeBias} object}

\item{print_statistics}{Whether to print calibration slope in the plot
legend.}
}
\description{
Private function to plot calibration curves
}
