% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_plot.R
\name{plot_roc}
\alias{plot_roc}
\title{Private function to plot ROC curves}
\usage{
plot_roc(x, print_statistics)
}
\arguments{
\item{x}{\code{seeBias} object}

\item{print_statistics}{Whether to print statistics in the plot legend.}
}
\description{
Private function to plot ROC curves
}
