% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_color_fill.R
\name{select_scale}
\alias{select_scale}
\title{#' Select appropriate color/fill scale based on number of sensitive groups
#' @param x seeBias object.
#' @param type Type to return, either "color" or "fill".
select_scale_old <- function(x, type = "color") {
  type <- match.arg(arg = type, choices = c("color", "fill"))
  n_sens <- x$input$n_sens
  if (n_sens > 6) {
    if (type == "color") scale_color_discrete else scale_fill_discrete
  } else {
    if (type == "color") scale_color_branded else scale_fill_branded
  }
}
Select appropriate color/fill scale based on number of sensitive groups}
\usage{
select_scale(x, type = "color")
}
\arguments{
\item{x}{seeBias object.}

\item{type}{Type to return, either "color" or "fill".}
}
\description{
#' Select appropriate color/fill scale based on number of sensitive groups
#' @param x seeBias object.
#' @param type Type to return, either "color" or "fill".
select_scale_old <- function(x, type = "color") {
  type <- match.arg(arg = type, choices = c("color", "fill"))
  n_sens <- x$input$n_sens
  if (n_sens > 6) {
    if (type == "color") scale_color_discrete else scale_fill_discrete
  } else {
    if (type == "color") scale_color_branded else scale_fill_branded
  }
}
Select appropriate color/fill scale based on number of sensitive groups
}
